Cache OAuth token in ~/.cache/meet/token.json
When the app runs, it checks whether there's a token there and if it's
still valid (i.e. hasn't expired). If it's still valid it reuses it.

The token is stored with a refresh token, which in principle allows you
to reobtain a new token even after it's expired. This bit hasn't been
implemented yet.